refactor: move Browse page
